+++ Changelog.md
@@ -30,6 +30,8 @@ Fixes:
   `xmlns:prefix=""`)
 - Fix DomReader's handling of non-namespace aware elements/nodes that return 
   `null` as localName.
+- In DomReader handle the fact that JVM's xpath creates xmlns declarations for
+  xmlns.
 
 # 0.85.0 – Tying things up
 *(Feb 19, 2023)<br />*